modules/preprocessing/downloaders/platform_downloader.py
#!/usr/bin/env python3
"""
平台下载器
处理平台特定的下载逻辑
"""

import os
import logging
from typing import Dict
from dotenv import load_dotenv

class PlatformDownloader:
    """平台下载器"""
    
    def __init__(self):
        """初始化平台下载器"""
        # 加载配置
        config_path = os.path.join(os.path.dirname(__file__), '..', '..', '..', 'config.env')
        load_dotenv(config_path)
        
        # 设置日志
        log_dir = os.path.join(os.path.dirname(__file__), '..', '..', '..', 'logs')
        os.makedirs(log_dir, exist_ok=True)
        
        logging.basicConfig(
            level=logging.INFO,
            format='%(asctime)s - PLATFORM_DOWNLOADER - %(levelname)s - %(message)s',
            handlers=[
                logging.FileHandler(os.path.join(log_dir, 'platform_downloader.log'), encoding='utf-8'),
                logging.StreamHandler()
            ]
        )
        
        self.logger = logging.getLogger('PlatformDownloader')
        self.logger.info("平台下载器已初始化")
    
    def download_from_platform(self, platform: str, url: str, **kwargs) -> Dict:
        """
        从平台下载文件
        :param platform: 平台名称
        :param url: 下载链接
        :param kwargs: 其他参数
        :return: 下载结果
        """
        try:
            self.logger.info(f"开始从{platform}平台下载: {url}")
            
            # 这里应该实现平台特定的下载逻辑
            # 目前返回基础信息
            return {
                "success": True,
                "platform": platform,
                "url": url,
                "message": f"平台{platform}下载功能待实现"
            }
            
        except Exception as e:
            self.logger.error(f"平台下载失败: {e}")
            return {
                "success": False,
                "error": str(e),
                "platform": platform,
                "url": url
            } 