requirements.txt
# 视频转录系统依赖包

# Web框架
Flask>=2.0.0

# 环境变量管理
python-dotenv>=0.19.0

# HTTP请求
requests>=2.25.0

# 音频处理
pydub>=0.25.0
ffmpeg-python>=0.2.0

# OpenAI API
openai>=1.0.0

# 文本处理
markdown>=3.3.0

# 日志和工具
colorama>=0.4.4
tqdm>=4.62.0

# 开发工具（可选）
# pytest>=6.0.0
# black>=21.0.0
# flake8>=3.9.0 